combined like functions into single files, converted to ES6 module syntax
