+++ src/box-box.js
@@ -1,5 +1,3 @@
-'use strict'
-
 /**
  * box-box collision
  * @param {number} x1 top-left corner of first box
@@ -11,7 +9,7 @@
  * @param {number} w2 width of second box
  * @param {number} h2 height of second box
  */
-module.exports = function boxBox(x1, y1, w1, h1, x2, y2, w2, h2)
+export function boxBox(x1, y1, w1, h1, x2, y2, w2, h2)
 {
     return x1 < x2 + w2 && x1 + w1 > x2 && y1 < y2 + h2 && y1 + h1 > y2
 }
