+++ src/box-point.js
@@ -1,5 +1,3 @@
-'use strict'
-
 /**
  * box-point collision
  * @param {number} x1 top-left corner of box
@@ -10,7 +8,22 @@
  * @param {number} y2 of point
  * @return {boolean}
  */
-module.exports = function boxPoint(x1, y1, w1, h1, x2, y2)
+export function boxPoint(x1, y1, w1, h1, x2, y2)
 {
     return x2 >= x1 && x2 <= x1 + w1 && y2 >= y1 && y2 <= y1 + h1
+}
+
+/**
+ * point-box collision
+ * @param {number} x1 point
+ * @param {number} y1 point
+ * @param {number} xb top-left corner of box
+ * @param {number} yb top-left corner of box
+ * @param {number} wb width of box
+ * @param {number} hb height of box
+ * @return {boolean}
+ */
+export function pointBox(x1, y1, xb, yb, wb, hb)
+{
+    return boxPoint(xb, yb, wb, hb, x1, y1)
 }
