--- src/box.js
@@ -1,8 +1,0 @@
-module.exports = {
-    point: require('./box-point'),
-    box: require('./box-box'),
-    line: require('./box-line'),
-    polygon: require('./box-polygon'),
-    circle: require('./box-circle'),
-    ellipse: require('./box-ellipse')
-}
