+++ src/circle-circleOutline.js
@@ -1,6 +1,15 @@
-var circleOutlineCircle = require('./circleOutline-circle')
+import {circlePoint} from './circle-point';
 
-module.exports = function circleCircleOutline(xco, yco, rco, xc, yc, rc)
+export function circleOutlineCircle(xc, yc, rc, xco, yco, rco, thickness)
+{
+    if (circlePoint(xc, yc, rc, xco, yco))
+    {
+        return true
+    }
+
+}
+
+export function circleCircleOutline(xco, yco, rco, xc, yc, rc)
 {
     return circleOutlineCircle(xc, yc, rc, xco, yco, rco)
 }
