--- src/circle.js
@@ -1,9 +1,0 @@
-module.exports = {
-    point: require('./circle-point'),
-    circle: require('./circle-circle'),
-    ellipse: require('./circle-ellipse'),
-    line: require('./circle-line'),
-    box: require('./circle-box'),
-    polygon: require('./circle-polygon'),
-    // circleOutline: require('./circle-circleOutline')
-}
