--- src/circleOutline-circle.js
@@ -1,10 +1,0 @@
-var circlePoint = require('./circle-point')
-
-module.exports = function circleOutlineCircle(xc, yc, rc, xco, yco, rco, thickness)
-{
-    if (circlePoint(xc, yc, rc, xco, yco))
-    {
-        return true
-    }
-
-}
