--- src/circleOutline.js
@@ -1,8 +1,0 @@
-module.exports = {
-    point: require('./circleOutline-point'),
-    circle: require('./circleOutline-circle'),
-    // ellipse: require('./circleOutline-ellipse'),
-    line: require('./circleOutline-line'),
-    box: require('./circleOutline-box'),
-    // polygon: require('./circleOutline-polygon')
-}
