+++ src/ellipse-line.js
@@ -10,7 +10,7 @@
  * @param {number} x2 second point of line
  * @param {number} y2 second point of line
  */
-module.exports = function ellipseLine(xe, ye, rex, rey, x1, y1, x2, y2)
+export function ellipseLine(xe, ye, rex, rey, x1, y1, x2, y2)
 {
     x1 -= xe
     x2 -= xe
@@ -37,4 +37,20 @@ module.exports = function ellipseLine(xe, ye, rex, rey, x1, y1, x2, y2)
     {
         return false
     }
+}
+
+/**
+ * line-ellipse collision
+ * @param {number} x1 first point of line
+ * @param {number} y1 first point of line
+ * @param {number} x2 second point of line
+ * @param {number} y2 second point of line
+ * @param {number} xe center of ellipse
+ * @param {number} ye center of ellipse
+ * @param {number} rx radius-x of ellipse
+ * @param {number} ry radius-y of ellipse
+ */
+export function lineEllipse(x1, y1, x2, y2, xe, ye, rex, rey)
+{
+    return ellipseLine(xe, ye, rex, rey, x1, y1, x2, y2)
 }
