--- src/ellipse.js
@@ -1,8 +1,0 @@
-module.exports = {
-    point: require('./ellipse-point'),
-    circle: require('./ellipse-circle'),
-    ellipse: require('./ellipse-ellipse'),
-    line: require('./ellipse-line'),
-    box: require('./ellipse-box'),
-    polygon: require('./ellipse-polygon')
-}
