--- src/index.js
@@ -1,51 +1,0 @@
-module.exports = {
-    circlePoint: require('./circle-point'),
-    circleCircle: require('./circle-circle'),
-    circleLine: require('./circle-line'),
-    circleBox: require('./circle-box'),
-    circlePolygon: require('./circle-polygon'),
-    circleEllipse: require('./circle-ellipse'),
-    // circleCircleOutline: require('./circle-circleOutline'),
-
-    circleOutlineBox: require('./circleOutline-box'),
-    circleOutlineLine: require('./circleOutline-line'),
-    circleOutlinePoint: require('./circleOutline-point'),
-    // circleOutlineCircle: require('./circleOutline-circle'),
-
-    polygonPoint: require('./polygon-point'),
-    polygonLine: require('./polygon-line'),
-    polygonPolygon: require('./polygon-polygon'),
-    polygonBox: require('./polygon-box'),
-    polygonCircle: require('./polygon-circle'),
-    polygonEllipse: require('./polygon-ellipse'),
-
-    boxPoint: require('./box-point'),
-    boxBox: require('./box-box'),
-    boxLine: require('./box-line'),
-    boxPolygon: require('./box-polygon'),
-    boxCircle: require('./box-circle'),
-    boxEllipse: require('./box-ellipse'),
-    boxCircleOutline: require('./box-circleOutline'),
-
-    pointBox: require('./point-box'),
-    pointPolygon: require('./point-polygon'),
-    pointCircle: require('./point-circle'),
-    pointLine: require('./point-line'),
-    pointEllipse: require('./point-ellipse'),
-    pointCircleOutline: require('./point-circleOutline'),
-
-    lineLine: require('./line-line'),
-    lineBox: require('./line-box'),
-    linePolygon: require('./line-polygon'),
-    lineCircle: require('./line-circle'),
-    linePoint: require('./line-point'),
-    lineEllipse: require('./line-ellipse'),
-    lineCircleOutline: require('./line-circleOutline'),
-
-    ellipsePoint: require('./ellipse-point'),
-    ellipseLine: require('./ellipse-line'),
-    ellipseBox: require('./ellipse-box'),
-    ellipseCircle: require('./ellipse-circle'),
-    ellipseEllipse: require('./ellipse-ellipse'),
-    ellipsePolygon: require('./ellipse-polygon')
-}
