+++ src/line-point.js
@@ -1,10 +1,3 @@
-'use strict'
-
-function distanceSquared(x1, y1, x2, y2)
-{
-    return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2))
-}
-
 /**
  * line-point collision
  * from https://stackoverflow.com/a/17693146/1955997
@@ -17,8 +10,27 @@ function distanceSquared(x1, y1, x2, y2)
  * @param {number} [tolerance=1]
  * @return {boolean}
  */
-module.exports = function linePoint(x1, y1, x2, y2, xp, yp, tolerance)
+export function linePoint(x1, y1, x2, y2, xp, yp, tolerance)
 {
     tolerance = tolerance || 1
     return Math.abs(distanceSquared(x1, y1, x2, y2) - (distanceSquared(x1, y1, xp, yp) + distanceSquared(x2, y2, xp, yp))) <= tolerance
+}
+function distanceSquared(x1, y1, x2, y2)
+{
+    return Math.sqrt(Math.pow(x1 - x2, 2) + Math.pow(y1 - y2, 2))
+}
+
+/**
+ * point-line collision
+ * @param {number} xp point
+ * @param {number} yp point
+ * @param {number} x1 first point in line
+ * @param {number} y1 first point in line
+ * @param {number} x2 second point in line
+ * @param {number} y2 second point in line
+ * @return {boolean}
+ */
+export function pointLine(xp, yp, x1, y1, x2, y2)
+{
+    return linePoint(x1, y1, x2, y2, xp, yp)
 }
