--- src/line.js
@@ -1,8 +1,0 @@
-module.exports = {
-    line: require('./line-line'),
-    box: require('./line-box'),
-    polygon: require('./line-polygon'),
-    circle: require('./line-circle'),
-    point: require('./line-point'),
-    ellipse: require('./line-ellipse')
-}
