--- src/point.js
@@ -1,7 +1,0 @@
-module.exports = {
-    box: require('./point-box'),
-    polygon: require('./point-polygon'),
-    circle: require('./point-circle'),
-    line: require('./point-line'),
-    ellipse: require('./point-ellipse')
-}
