--- src/polygon.js
@@ -1,9 +1,0 @@
-'use strict'
-
-module.exports = {
-    point: require('./polygon-point'),
-    line: require('./polygon-line'),
-    polygon: require('./polygon-polygon'),
-    box: require('./polygon-box'),
-    circle: require('./polygon-circle')
-}
