+++ src/umd-index.js
@@ -1,0 +1,62 @@
+//box-function imports
+import {boxPoint} from './box-point';
+import {boxBox} from './box-box';
+import {boxLine} from './box-line';
+import {boxPolygon} from './box-polygon';
+import {boxCircle} from './box-circle';
+import {boxEllipse} from './box-ellipse';
+import {boxCircleOutline} from './box-circleOutline';
+
+//circle-function imports
+import {circlePoint} from './circle-point';
+import {circleCircle} from './circle-circle';
+import {circleLine} from './circle-line';
+import {circleBox} from './box-circle';
+import {circlePolygon} from './circle-polygon';
+import {circleEllipse} from './circle-ellipse';
+import {circleCircleOutline} from './circle-circleOutline';
+
+//circleOutline-function imports
+import {circleOutlineBox} from './box-circleOutline';
+import {circleOutlineLine} from './circleOutline-line';
+import {circleOutlinePoint} from './circleOutline-point';
+import {circleOutlineCircle} from './circle-circleOutline';
+
+//ellipse-function imports
+import {ellipsePoint} from './ellipse-point';
+import {ellipseLine} from './ellipse-line';
+import {ellipseBox} from './box-ellipse';
+import {ellipseCircle} from './circle-ellipse';
+import {ellipseEllipse} from './ellipse-ellipse';
+import {ellipsePolygon} from './ellipse-polygon';
+
+//line-function imports
+import {lineBox} from './box-line';
+import {lineCircle} from './circle-line';
+import {lineCircleOutline} from './circleOutline-line';
+import {lineEllipse} from './ellipse-line';
+import {lineLine} from './line-line';
+import {linePolygon} from './line-polygon';
+import {linePoint} from './line-point';
+
+//polygon-function imports
+import {polygonBox} from './box-polygon';
+import {polygonCircle} from './circle-polygon';
+import {polygonEllipse} from './ellipse-polygon';
+import {polygonLine} from './line-polygon';
+import {polygonPolygon} from './polygon-polygon';
+import {polygonPoint} from './polygon-point';
+
+//point-function imports
+import {pointBox} from './box-point';
+import {pointCircle} from './circle-point';
+import {pointCircleOutline} from './circleOutline-point';
+import {pointEllipse} from './ellipse-point';
+import {pointLine} from './line-point';
+import {pointPolygon} from './polygon-point';
+
+//handle exports
+export * as Intersects from '...';
+
+
+
